requirements.txt
# Core CrewAI dependencies
crewai>=0.28.0
crewai-tools>=0.1.6

# LangChain and AI providers
langchain>=0.1.0
openai>=1.0.0

# Optional: Additional AI providers
anthropic>=0.7.0
google-generativeai>=0.3.0

# Development tools
python-dotenv>=1.0.0
jupyter>=1.0.0
ipython>=8.0.0

# Testing and quality
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Utilities
rich>=13.0.0
typer>=0.9.0
